requirements.txt
langchain>=0.3.19
langchain-community>=0.3.18
langchain-core>=0.3.39
langchain-groq>=0.2.4
langchain-google-genai>=2.0.10
langgraph>=0.2.74
pydantic>=2.10.6
python-dotenv>=1.0.1
qdrant-client>=1.13.2
sentence-transformers>=3.4.1
huggingface-hub>=0.29.1
replicate>=1.0.4
tavily-python>=0.5.1
numpy>=2.2.3
tenacity>=9.0.0
pypdf
docx2txt
unstructured 